requirements.txt
fastapi==0.110.1
uvicorn==0.29.0

gspread==5.12.4
oauth2client==4.1.3

jinja2==3.1.3
python-docx==1.1.0
docx2pdf==0.1.8  # Optional: only if you need to convert .docx to .pdf

python-dotenv==1.0.1
